# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "InvoiceTool Swagger API",
        "description": "InvoiceTool Swagger API description",
        "contact": {
            "email": "admin@finpay.com"
        },
        "version": "1.0.0"
    },
    "host": "invoice-tool-nov8.test/",
    "basePath": "api/v1",
    "paths": {
...
        "/Currencies": {
            "get": {
                "tags": [
                    "CURRENCIES"
                ],
                "summary": "Currencies Table",
                "operationId": "cid",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/currencies": {
            "post": {
                "tags": [
                    "CURRENCIES"
                ],
                "summary": "Currencies Store",
                "description": "Currencies Store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "placement",
                        "in": "formData",
                        "description": "Symbol Placement",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "code",
                        "in": "formData",
                        "description": "Code",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "decimal",
                        "in": "formData",
                        "description": "Decimal Point",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "symbol",
                        "in": "formData",
                        "description": "Symbol",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "thousands",
                        "in": "formData",
                        "description": "Thousands Separator",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/Currencies/{id}": {
            "get": {
                "tags": [
                    "CURRENCIES"
                ],
                "summary": "Currencies edit",
                "operationId": "Ceid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currencies Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CURRENCIES"
                ],
                "summary": "Currencies Update",
                "description": "Currencies Update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currencies Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "placement",
                        "in": "formData",
                        "description": "Symbol Placement",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "code",
                        "in": "formData",
                        "description": "Code",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "decimal",
                        "in": "formData",
                        "description": "Decimal Point",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "symbol",
                        "in": "formData",
                        "description": "Symbol",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "thousands",
                        "in": "formData",
                        "description": "Thousands Separator",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CURRENCIES"
                ],
                "summary": "Currencies Delete",
                "operationId": "Cdid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currencies Delete",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "auth number",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
}